Adding a new initialization plugin RemoveCells that removes cells that were created but not necessary. For example, when only a circular domain is needed but a rectangular domain of cells has been created by earlier initializers.
I first tried to do this by inheriting 2 interfaces in the CellDeath plugin such that it can be used as an Initializer. Although this basically works and it scheduled at init time, it will be scheduled ALSO as a cell type plugin (which it should not). Therefore, I decided to create a new plugin for this purpose (while reducing redundant functionality, such as Shrinkage).