SymbolAccessor mode that allows partial symbol definition
For plot generation it is convenient to plot expressions/symbols just for the contexts in which they are defined and ignore all others. The allowPartialSpec-option is designed to jump in here. SymbolAccessors can be created also for contexts in which they are partially specified. Accessors will throw SymbolError of type SymbolError::Type::Undefined when the selected focus does not define the required symbol(s).